Commit transform int to float finished.

+++ src/GameObjects/FasterEnemy.java
@@ -34,11 +34,11 @@ public class FasterEnemy extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.CYAN);
-        g.fillRect(x,y,WIDTH,HEIGHT);
+        g.fillRect((int)x,(int)y,WIDTH,HEIGHT);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x,y,WIDTH,HEIGHT);
+        return new Rectangle((int)x,(int)y,WIDTH,HEIGHT);
     }
 }
